requirements.txt
langchain==0.1.9
streamlit==1.31.0
pypdf2==3.0.1
python-dotenv==1.0.1